Successfully integrate  Chat,  Share Files and Face.
Significant changes:
Everytime I switch the fragment, I will close the interestId, registerPrefixId, face.

In the function "Share Files", if one send the same Interest, since its CS(content store) has remember the data packet for this interest, it NFD will not send this interest to the requested face. That's why the the GI of the other phone(content provider) doesn't change.

Next Step:
Add the "Rrfresh" to get the new content of the content provider.